Reorganize files a bit more
