+++ src/__tests__/object.test.js
@@ -1,6 +1,6 @@
 // @flow strict
 
-import * as Result from '../lib/Result';
+import * as Result from '../core/Result';
 import { exact, inexact, object, pojo } from '../object';
 import { guard } from '../guard';
 import { hardcoded } from '../constants';
