+++ src/core/Result.js
