+++ src/core/__tests__/Annotation.test.js
