+++ src/core/__tests__/Result.test.js
