+++ src/types/core/Result.d.ts
