+++ src/types/format/index.d.ts
@@ -1,4 +1,4 @@
-import { Annotation } from '../Annotation';
+import { Annotation } from '../core/Annotation';
 
 export function serialize(ann: Annotation): string;
 export function summarize(ann: Annotation): string;
