+++ src/types/json-tests.ts
@@ -1,5 +1,5 @@
 import { json, jsonObject, jsonArray } from 'decoders';
-import * as Result from 'decoders/lib/Result';
+import * as Result from 'decoders/core/Result';
 
 // $ExpectType JSONValue
 Result.unwrap(json('hi'));
